User interface and server implementation completed

+++ src/Server/Service/MovieTicket.java
@@ -10,13 +10,8 @@ import java.rmi.RemoteException;
 import java.rmi.server.UnicastRemoteObject;
 import java.text.ParseException;
 import java.util.*;
-import java.util.concurrent.ConcurrentHashMap;
 
 public class MovieTicket extends UnicastRemoteObject implements IMovieTicket{
-    // MovieName,MovieID,BookingCapacity
-    private Map<String, Map<String, MovieState>> movies;
-    //CustomerID,MovieID,BookingCapacity
-    private Map<String, Map<String, User>> customers;
     private IServerInfo serverInfo;
     private IUdp udpService;
     private IMovie movieService;
@@ -28,8 +23,6 @@ public class MovieTicket extends UnicastRemoteObject implements IMovieTicket{
                        ICustomerBooking customerBookingDb,
                        IMovies moviesDb) throws RemoteException {
         super();
-        this.movies = new ConcurrentHashMap<>();
-        this.customers = new ConcurrentHashMap<>();
         this.serverInfo = serverInfo;
         this.udpService = udpService;
         this.movieService = movieService;
@@ -48,48 +41,58 @@ public class MovieTicket extends UnicastRemoteObject implements IMovieTicket{
     }
 
     public String removeMovieSlots(String movieId, String movieName) throws RemoteException, ParseException {
-        if (this.moviesDb.ifMovieNameExist(movieName)) {
-            if (this.moviesDb.ifMovieIDExist(movieName, movieId)) {
-                //TODO search if customer booked the ticket that is going to be deleted by Admin
-                List<String> bookingCustomerIDs = this.customerBookingDb.getAllCustomerIDs();
-                String bookingCustomerID = null;
-                for (int i = 0; i < bookingCustomerIDs.size(); i++) {
-                    String customerID = bookingCustomerIDs.get(0);
-                    Map<String, MovieState> bookings = (Map<String, MovieState>) this.customerBookingDb.getTicketsBookedByCustomerID(customerID);
-                    if (bookings.get(movieId) != null) {
-                        for (int j = 0; j < bookings.size(); j++) {
-                            if (bookings.get(0).getMovieName().equals(movieName) && bookings.get(movieId).equals(movieId)) {
+        if (!Util.getServerPrefixByMovieID(movieId).equals(this.serverInfo.getServerName())) {
+            this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(Util.getServerPrefixByMovieID(movieId)), "removeMovieSlots", null, movieName, movieId, -1);
+        } else {
+            if (this.moviesDb.ifMovieNameExist(movieName)) {
+                if (this.moviesDb.ifMovieIDExist(movieName, movieId)) {
+                    List<String> bookingCustomerIDs = this.customerBookingDb.getAllCustomerIDs();
+                    String bookingCustomerID = "";
+                    for (String customerID : bookingCustomerIDs) {
+                        Map<String, MovieState> bookings = (Map<String, MovieState>) this.customerBookingDb.getTicketsBookedByCustomerID(customerID);
+                        if (bookings.get(movieId) != null) {
+                            if (bookings.get(movieId).getMovieTicketInfo().get(movieName) != null) {
                                 bookingCustomerID = customerID;
                                 break;
                             }
                         }
                     }
-                }
-                if (bookingCustomerID != null) {
-                    List<String> availableMovieSlots = this.moviesDb.getMovieSlotsAtSpecificArea(movieName, this.serverInfo.getServerName());
-                    if (!availableMovieSlots.isEmpty()) {
-                        List<MovieState> movieInfo = new ArrayList<MovieState>();
-                        for (String availableMovieSlot : availableMovieSlots) {
-                            movieInfo.add(new MovieState(movieName, String.valueOf(availableMovieSlot), 0));
-                        }
-
-                        movieInfo.sort(new Comparator<MovieState>() {
-                            public int compare(MovieState movieA, MovieState movieB) {
-                                return movieA.getMovieDate().compareTo(movieB.getMovieDate());
+                    if (!bookingCustomerID.isEmpty()) {
+                        String nextAvailableBookingID = this.findNextAvailableSlot(bookingCustomerID,movieId,movieName);
+                        if(!nextAvailableBookingID.isEmpty()) {
+                            int currentNumberOfTicketBookedByCustomer = this.customerBookingDb.getNoOfTicketsBookedByMovieID(bookingCustomerID, movieId, movieName);
+                            this.customerBookingDb.addMovieByCustomerID(bookingCustomerID, nextAvailableBookingID, movieName, currentNumberOfTicketBookedByCustomer);
+                            this.customerBookingDb.cancelMovieByMovieID(bookingCustomerID, movieId,movieName);
+                        } else {
+                            String nextAvailableMovieIDAtwater = "";
+                            String nextAvailableMovieIDVerdun = "";
+                            String nextAvailableMovieIDOutermont = "";
+                            if(!this.serverInfo.getServerName().equals(ServerConstant.SERVER_ATWATER_PREFIX)) {
+                                nextAvailableMovieIDAtwater = this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(Util.getServerPrefixByMovieID(movieId)), "findNextAvailableSlot", bookingCustomerID, movieName, movieId, -1);
                             }
-                        });
-
-                        for (int i = 0; i < movieInfo.size(); i++) {
-                            if (i + 1 < movieInfo.size() && movieInfo.get(0).getMovieID().equals(movieId)) {
-                                int noOfTicket = this.customerBookingDb.getNoOfTicketsBookedByMovieID(bookingCustomerID, movieId, movieName);
-                                this.customerBookingDb.addMovieByCustomerID(bookingCustomerID, movieInfo.get(i + 1).getMovieID(), movieName, noOfTicket);
-                                break;
+                            if(!this.serverInfo.getServerName().equals(ServerConstant.SERVER_VERDUN_PREFIX)) {
+                                nextAvailableMovieIDVerdun = this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(Util.getServerPrefixByMovieID(movieId)), "findNextAvailableSlot", bookingCustomerID, movieName, movieId, -1);
+                            }
+                            if(!this.serverInfo.getServerName().equals(ServerConstant.SERVER_OUTREMONT_PREFIX)) {
+                                nextAvailableMovieIDOutermont = this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(Util.getServerPrefixByMovieID(movieId)), "findNextAvailableSlot", bookingCustomerID, movieName, movieId, -1);
+                            }
+                            List<MovieState> movieInfo = new ArrayList<MovieState>();
+                            movieInfo.add(new MovieState(movieName, String.valueOf(nextAvailableMovieIDAtwater), 0));
+                            movieInfo.add(new MovieState(movieName, String.valueOf(nextAvailableMovieIDVerdun), 0));
+                            movieInfo.add(new MovieState(movieName, String.valueOf(nextAvailableMovieIDOutermont), 0));
+                            List<MovieState> sortedList = Util.sortMovieByDates(movieInfo);
+                            if(sortedList.size()>0) {
+                                int currentNumberOfTicketBookedByCustomer = this.customerBookingDb.getNoOfTicketsBookedByMovieID(bookingCustomerID, movieId, movieName);
+                                this.customerBookingDb.addMovieByCustomerID(bookingCustomerID, sortedList.get(0).getMovieID(), movieName, currentNumberOfTicketBookedByCustomer);
+                                this.customerBookingDb.cancelMovieByMovieID(bookingCustomerID, movieId,movieName);
+                                this.moviesDb.decrementBookingCapacity(movieName,movieId,currentNumberOfTicketBookedByCustomer);
+                                this.moviesDb.deleteMovieSlotByMovieNameAndMovieID(movieName, movieId);
+                                return "Next Slot Booked and Slot deleted from theater";
                             }
                         }
                     }
-                    this.customerBookingDb.cancelMovieByMovieID(bookingCustomerID, movieId);
+                    return this.moviesDb.deleteMovieSlotByMovieNameAndMovieID(movieName, movieId);
                 }
-                return this.moviesDb.deleteMovieSlotByMovieNameAndMovieID(movieName, movieId);
             }
         }
         return "Movie does not exist";
@@ -98,30 +101,52 @@ public class MovieTicket extends UnicastRemoteObject implements IMovieTicket{
     public String listMovieShowsAvailability(String movieName) throws RemoteException {
         StringBuilder sb = new StringBuilder();
         sb.append(this.getMoviesListInTheatre(movieName));
-        System.out.println("Calling Server 1");
         if(!this.serverInfo.getServerName().equals(ServerConstant.SERVER_ATWATER_PREFIX)) sb.append(this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(ServerConstant.SERVER_ATWATER_PREFIX),"getMoviesListInTheatre",null,movieName,null,-1));
-        System.out.println("Calling Server 2");
         if(!this.serverInfo.getServerName().equals(ServerConstant.SERVER_VERDUN_PREFIX)) sb.append(this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(ServerConstant.SERVER_VERDUN_PREFIX),"getMoviesListInTheatre",null,movieName,null,-1));
-        System.out.println("Calling Server 3");
         if(!this.serverInfo.getServerName().equals(ServerConstant.SERVER_OUTREMONT_PREFIX)) sb.append(this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(ServerConstant.SERVER_OUTREMONT_PREFIX),"getMoviesListInTheatre",null,movieName,null,-1));
         return sb.toString();
     }
 
     public String bookMovieTickets(String customerID, String movieId, String movieName, int numberOfTickets) throws RemoteException, ParseException {
-        String movieInTheater = this.movieService.grepServerPrefixByMovieID(movieId);
-        if(!movieInTheater.equals(this.serverInfo.getServerName())) return this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(movieInTheater),"bookTicket",customerID,movieName,movieId,numberOfTickets);
-        return this.bookTicket(customerID,movieId,movieName,numberOfTickets);
+        StringBuilder sb = new StringBuilder();
+        if (!Util.getServerPrefixByMovieID(movieId).equals(this.serverInfo.getServerName())) {
+            if(!this.customerBookingDb.ifMovieBookingExist(customerID,movieId,movieName))
+                return this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(Util.getServerPrefixByMovieID(movieId)), "bookTicket", customerID, movieName, movieId, numberOfTickets);
+            else {
+                sb.append("Movie Already booked at ").append(Util.getServerFullNameByCustomerID(customerID)).append(" theater");
+                return sb.toString();
+            }
+        }
+        if(numberOfTickets<this.moviesDb.getSlotBookingCapacity(movieName,movieId)) {
+            return this.bookTicket(customerID, movieId, movieName, numberOfTickets,true);
+        }
+        sb.append("Seats cannot be booked for more than ").append(this.moviesDb.getSlotBookingCapacity(movieName,movieId)).append(" customers");
+        return sb.toString();
     }
 
     public String getBookingSchedule(String customerID) throws RemoteException {
-        String customerRegisteredToServer = customerID.substring(0,3);
-        if(!customerRegisteredToServer.equals(this.serverInfo.getServerName())) return this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(customerRegisteredToServer),"getCustomerBookingList",customerID,null,null,-1);
-        return this.getCustomerBookingList(customerID);
+        StringBuilder sb = new StringBuilder();
+        sb.append(Util.getServerFullNameByCustomerID(customerID)).append("\n");
+        sb.append(this.getCustomerBookingList(customerID));
+        if(!this.serverInfo.getServerName().equals(ServerConstant.SERVER_ATWATER_PREFIX)) {
+            sb.append("ATWATER \n");
+            sb.append(this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(ServerConstant.SERVER_ATWATER_PREFIX),"getCustomerBookingList",customerID,null,null,-1));
+        }
+        if(!this.serverInfo.getServerName().equals(ServerConstant.SERVER_VERDUN_PREFIX)) {
+            sb.append("VERDUN \n");
+            sb.append(this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(ServerConstant.SERVER_VERDUN_PREFIX),"getCustomerBookingList",customerID,null,null,-1));
+        }
+        if(!this.serverInfo.getServerName().equals(ServerConstant.SERVER_OUTREMONT_PREFIX)) {
+            sb.append("OUTERMONT \n");
+            sb.append(this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(ServerConstant.SERVER_OUTREMONT_PREFIX),"getCustomerBookingList",customerID,null,null,-1));
+        }
+        return sb.toString();
     }
 
     public String cancelMovieTickets(String customerID, String movieID, String movieName, int numberOfTickets) throws RemoteException {
-        String customerRegisteredToServer = customerID.substring(0,3);
-        if(!customerRegisteredToServer.equals(this.serverInfo.getServerName())) return this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(customerRegisteredToServer),"cancelTicket",customerID,movieName,movieID,numberOfTickets);
+        if (!Util.getServerPrefixByMovieID(movieID).equals(this.serverInfo.getServerName())) {
+            return this.udpService.sendUDPMessage(this.serverInfo.getServerPortNumber(Util.getServerPrefixByMovieID(movieID)),"cancelTicket",customerID,movieName,movieID,numberOfTickets);
+        }
         return cancelTicket(customerID, movieID, movieName, numberOfTickets);
     }
 
@@ -129,24 +154,45 @@ public class MovieTicket extends UnicastRemoteObject implements IMovieTicket{
         Map<String, Integer> movieSlots = this.moviesDb.getMovieSlotsHashMapByMovieName(movieName);
         if(movieSlots!=null) {
             StringBuilder builder = new StringBuilder();
+            builder.append(Util.getServerNameByServerPrefix(this.serverInfo.getServerName())).append(" \n");
             for (Map.Entry<String,Integer> slot : movieSlots.entrySet()) {
-                builder.append(movieName + ": " +slot.getValue() +" "+ slot.getKey()+", ");
+                builder.append("Movie Name: "+ movieName + " |" + " Seats Available: "+ slot.getValue()+ "| Slot: " + Util.getSlotByMovieID(slot.getKey()) + "| Movie Date: " + Util.getSlotDateByMovieID(slot.getKey()) + ", \n\n");
             }
             return builder.toString();
         }
         return "";
     }
 
-    public String bookTicket(String customerID, String movieId, String movieName, int numberOfTickets) throws RemoteException, ParseException {
+    public String bookTicket(String customerID, String movieId, String movieName, int numberOfTickets, boolean isUserRegisteredToServer) throws RemoteException, ParseException {
         StringBuilder builder = new StringBuilder();
-        if(this.moviesDb.ifMovieIDExist(movieName,movieId)) {
-            this.customerBookingDb.addMovieByCustomerID(customerID,movieId,movieName,numberOfTickets);
-            builder.append("Movie Added Successfully");
-            builder.append(this.moviesDb.decrementBookingCapacity(movieName,movieId,numberOfTickets));
+        if(numberOfTickets<this.moviesDb.getSlotBookingCapacity(movieName,movieId)){
+            if(this.moviesDb.ifMovieIDExist(movieName,movieId)) {
+                if(isUserRegisteredToServer) {
+                    this.customerBookingDb.addMovieByCustomerID(customerID,movieId,movieName,numberOfTickets);
+                    builder.append("Movie Added Successfully \n");
+                    builder.append(this.moviesDb.decrementBookingCapacity(movieName,movieId,numberOfTickets));
+                }else {
+                    if(numberOfTickets<=this.moviesDb.getSlotBookingCapacity(movieName,movieId)) {
+                        if(this.customerBookingDb.noOfMoviesBookedInAWeek(customerID,movieId)<3) {
+                            this.customerBookingDb.addMovieByCustomerID(customerID,movieId,movieName,numberOfTickets);
+                            builder.append("Movie Added Successfully \n");
+                            builder.append(this.moviesDb.decrementBookingCapacity(movieName,movieId,numberOfTickets));
+                        }else {
+                            builder.append("You cannot book more than 3 movie in this theater\n");
+                        }
+                    }else {
+                        builder.append("Seats cannot be booked for more than ").append(this.moviesDb.getSlotBookingCapacity(movieName,movieId)).append(" customers");
+                    }
+                }
+                return builder.toString();
+            } else {
+                builder.append("Booking Failed");
+                return builder.toString();
+            }
+        } else {
+            builder.append("Booking slots filled");
             return builder.toString();
         }
-        builder.append("Booking Failed");
-        return builder.toString();
     }
 
     public String getCustomerBookingList(String customerID) throws RemoteException {
@@ -155,7 +201,9 @@ public class MovieTicket extends UnicastRemoteObject implements IMovieTicket{
             StringBuilder builder = new StringBuilder();
             for (MovieState bookingSchedule :
                     customerObj.values()) {
-                builder.append(bookingSchedule.getMovieName().toString() + ": " + bookingSchedule.getMovieID() + " " + bookingSchedule.getRemainingSlots());
+                bookingSchedule.getMovieTicketInfo().entrySet().forEach(entry -> {
+                    builder.append(entry.getKey()+ " Slot: " + Util.getSlotByMovieID(bookingSchedule.getMovieID()) + "| MovieID: " + bookingSchedule.getMovieID() + "| Seats Booked: " + entry.getValue() + ", \n");
+                });
             }
             return builder.toString();
         }
@@ -163,6 +211,28 @@ public class MovieTicket extends UnicastRemoteObject implements IMovieTicket{
     }
 
     public String cancelTicket(String customerID, String movieID, String movieName, int numberOfTickets) throws RemoteException {
+        this.moviesDb.incrementBookingCapacity(movieName,movieID,numberOfTickets);
         return this.customerBookingDb.cancelMovieTickets(customerID,movieID,movieName);
+    }
+
+    public String findNextAvailableSlot(String customerID,String movieID, String movieName) throws ParseException {
+        List<String> availableMovieSlots = this.moviesDb.getMovieSlotsAtSpecificArea(movieName, this.serverInfo.getServerName());
+        if (!availableMovieSlots.isEmpty()) {
+            List<MovieState> movieInfo = new ArrayList<MovieState>();
+            for (String availableMovieSlot : availableMovieSlots) {
+                movieInfo.add(new MovieState(movieName, String.valueOf(availableMovieSlot), 0));
+            }
+
+            movieInfo = Util.sortMovieBySlots(movieInfo);
+
+            String nextAvailableBookingID = "";
+            for (int j = 0; j < movieInfo.size(); j++) {
+                if(movieInfo.get(j).getMovieID().equals(movieID)) {
+                    if(j+1<movieInfo.size()) nextAvailableBookingID = movieInfo.get(j+1).getMovieID();
+                }
+            }
+            return nextAvailableBookingID;
+        }
+        return "";
     }
 }
